+++ Changes.md
@@ -3,7 +3,10 @@
 
 Improvements
 ------------
-- Spreadsheet : Enabled cell selection, editing is now achieved with a double-click.
+- Spreadsheet :
+  - Enabled cell selection, editing is now achieved with a double-click.
+  - Added menu items to copy/paste the enabled state and value(s) of selected cells.
+  - Added menu items to delete the selected rows.
 
 Fixes
 -----